requirements.txt
discord
python-dotenv
mysql-connector-python
